fix: Improve error handling
