+++ Cargo.toml
@@ -6,11 +6,11 @@ edition = "2018"
 
 [dependencies]
 alsa = "0"
-failure = "0"
 libnotify = "1.0"
 unixbar = { git = "https://github.com/agraven/unixbar" }
 pulsectl = { version = "0.2", package = "rust-pulsectl-fork" }
 systemstat = "0"
+anyhow = "1.0"
 
 [profile.release]
 lto = true
