+++ src/main.rs
@@ -18,6 +18,8 @@ extern crate failure;
 extern crate systemstat;
 extern crate unixbar;
 
+mod volume;
+
 use std::{
 	io::ErrorKind as IoErrorKind,
 	path::Path,
@@ -39,119 +41,18 @@ use unixbar::{
 	},
 	Duration, UnixBar,
 };
+use anyhow::{anyhow, Result};
 
-mod volume {
-	use std::{
-		io::Write,
-		process::{Command, Stdio},
-	};
 
-	use alsa::mixer::{Mixer, SelemChannelId, SelemId};
-	use libnotify::Notification;
-	use pulsectl::controllers::{AppControl, DeviceControl, SinkController};
-
-	pub fn add(diff: i8) -> alsa::Result<()> {
-		let mixer = Mixer::new("default", false)?;
-		let se_id = SelemId::new("Master", 0);
-		let selem = mixer.find_selem(&se_id).unwrap();
-		let (min, max) = selem.get_playback_volume_range();
-		// Current volume
-		let volume = selem.get_playback_volume(SelemChannelId::FrontLeft)?;
-		// A single percent volume
-		let step = (max - min) as f64 * 0.01;
-		let new_volume = volume + (step * f64::from(diff)).round() as i64;
-		selem.set_playback_volume_all(new_volume.max(min).min(max))?;
-		/*let _ = Command::new("pactl")
-		.arg("set-sink-volume")
-		.arg("@DEFAULT_SINK@")
-		.arg(format!("{:+}%", diff))
-		.spawn();**/
-		Ok(())
-	}
-
-	pub fn set_device(controller: &mut SinkController, name: &str) -> Result<(), failure::Error> {
-		// Set default device
-		match controller.set_default_device(name) {
-			Ok(false) => Notification::new("Couldn't set new device", None, None).show()?,
-			Err(e) => Notification::new(
-				"Error setting default device",
-				Some(format!("{:?}", e).as_str()),
-				None,
-			)
-			.show()?,
-			_ => (),
-		}
-		// Change output for all active streams
-		for app in controller.list_applications().unwrap() {
-			if let Err(e) = controller.move_app_by_name(app.index, name) {
-				Notification::new(
-					"Error changing sink for applicaiton",
-					Some(format!("{:?}", e).as_str()),
-					None,
-				)
-				.show()?;
-			}
-		}
-		Ok(())
-	}
-
-	pub fn menu() -> Result<(), failure::Error> {
-		let mut controller = SinkController::create()?;
-		// Launch device selection dialogue
-		let mut cmd = Command::new("zenity")
-			.args(&[
-				"--list",
-				"--text=Choose an audio device",
-				"--column=device-id",
-				"--column=Device name",
-				"--hide-column=1",
-				"--width=450",
-				"--height=250",
-			])
-			.stdin(Stdio::piped())
-			.stdout(Stdio::piped())
-			.spawn()?;
-		// Write device names to process stdin
-		{
-			let mut stdin = cmd.stdin.as_mut().unwrap();
-			for device in controller.list_devices().unwrap_or_default() {
-				writeln!(&mut stdin, "{}", device.name.unwrap_or_default())?;
-				writeln!(&mut stdin, "{}", device.description.unwrap_or_default())?;
-			}
-		}
-		// Get process stdout
-		let output = cmd.wait_with_output()?;
-		let new_device = String::from_utf8_lossy(&output.stdout);
-		let new_device = new_device.trim();
-		// Set audio device
-		if !new_device.is_empty() {
-			set_device(&mut controller, new_device)?;
-		}
-		Ok(())
-	}
-
-	/// Toggles whether volume is muted
-	pub fn mute() -> alsa::Result<()> {
-		let mixer = Mixer::new("default", false)?;
-		let se_id = SelemId::new("Master", 0);
-		let selem = mixer.find_selem(&se_id).unwrap();
-
-		let muted = selem.get_playback_switch(SelemChannelId::FrontLeft)? == 0;
-		selem.set_playback_switch_all(if muted { 1 } else { 0 })?;
-		Ok(())
-	}
-
-	pub fn icon(vol: u8) -> &'static str {
-		match vol {
-			0..=29 => "",
-			30..=59 => "",
-			_ => "",
-		}
-	}
+fn catch<F: FnMut() -> Result<Format, anyhow::Error>>(mut closure: F) -> Format {
+    match closure() {
+        Ok(fmt) => fmt,
+        Err(e) => bfmt![fg["#ff5555"] fmt["{}", e.to_string()]]
+    }
 }
 
-fn main() {
-	libnotify::init(env!("CARGO_PKG_NAME")).unwrap();
+fn main() -> Result<()> {
+	libnotify::init(env!("CARGO_PKG_NAME")).map_err(|e| anyhow!(e))?;
 	let battery_warned = Arc::new(AtomicBool::new(false));
 	// The structure representing the bar to generate
 	let formatter = I3BarFormatter::new();
@@ -216,16 +117,16 @@ fn main() {
 			bfmt![fmt["⌨ {}", layout]]
 		}))
 		// Disk space
-		.add(Periodic::new(Duration::from_secs(2), || {
+		.add(Periodic::new(Duration::from_secs(2), || catch(|| {
 			// Get the filesystem mounted at root
-			let fs = System::new().mount_at(Path::new("/")).unwrap();
-			bfmt![
+			let fs = System::new().mount_at(Path::new("/"))?;
+			Ok(bfmt![
 				fg["#cccccc"]
 				fmt[" {}", fs.avail.to_string()]
-			]
-		}))
+			])
+		})))
 		// Access point name
-		.add(Periodic::new(Duration::from_secs(1), || {
+		.add(Periodic::new(Duration::from_secs(1), || catch(|| {
 			let nmcli = |args: &[&str]| -> String {
 				Command::new("nmcli")
 					.args(args)
@@ -250,35 +151,35 @@ fn main() {
 					_ => "﹖",
 				})
 				.unwrap_or("﹖");
-			let (name, color) = match connection.split(':').next().unwrap() {
+			let (name, color) = match connection.split(':').next().ok_or_else(|| anyhow!("nmcli returned unexpected data!"))? {
 				"" => ("Disconnected", "#BB5555"),
 				name => (name, "#99ee99"),
 			};
-			bfmt![
+			Ok(bfmt![
 				fg[color]
 				fmt["{} {}{}", icon, name, status]
-			]
-		}))
+			])
+		})))
 		// Load average
-		.add(Periodic::new(Duration::from_secs(1), || {
-			let load = System::new().load_average().unwrap();
-			bfmt![
+		.add(Periodic::new(Duration::from_secs(1), || catch (|| {
+			let load = System::new().load_average()?;
+			Ok(bfmt![
 				fg["#cc9999"]
 				fmt[" {:.2}", load.one]
-			]
-		}))
+			])
+		})))
 		// Memory
-		.add(Periodic::new(Duration::from_secs(2), || {
-			let memory = System::new().memory().unwrap();
+		.add(Periodic::new(Duration::from_secs(2), || catch(|| {
+			let memory = System::new().memory()?;
 			let free = memory.free.as_u64() as f32 / 1_000_000_000.0;
-			bfmt![
+			Ok(bfmt![
 				fg["#ffc300"]
 				fmt[" {:.1} G", free]
-			]
-		}))
+			])
+		})))
 		// Temperature
-		.add(Periodic::new(Duration::from_secs(2), || {
-			let temp = System::new().cpu_temp().unwrap();
+		.add(Periodic::new(Duration::from_secs(2), || catch(|| {
+			let temp = System::new().cpu_temp()?;
 			let icon = match temp as u32 {
 				0..=59 => "",
 				60..=69 => "",
@@ -286,20 +187,20 @@ fn main() {
 				80..=89 => "",
 				_ => "",
 			};
-			bfmt![
+			Ok(bfmt![
 				fg["#10ff10"]
 				fmt["{} {:.1} °C", icon, temp]
-			]
-		}))
+			])
+		})))
 		// Battery
-		.add(Periodic::new(Duration::from_secs(1), move || {
+		.add(Periodic::new(Duration::from_secs(1), move || catch(|| {
 			let charging = match System::new().on_ac_power() {
 				Ok(on_ac) => on_ac,
-				_ => return bfmt![text[""]],
+				_ => return Ok(bfmt![text[""]]),
 			};
 			let battery = match System::new().battery_life() {
 				Ok(battery) => battery,
-				_ => return bfmt![text[""]],
+				_ => return Ok(bfmt![text[""]]),
 			};
 			let capacity = (battery.remaining_capacity * 100.0).round() as u8;
 
@@ -313,7 +214,7 @@ fn main() {
 						Some("battery-caution"),
 					);
 					notif.set_urgency(Urgency::Critical);
-					notif.show().unwrap();
+					notif.show()?;
 					battery_warned.store(true, Ordering::Release);
 				}
 			} else {
@@ -329,11 +230,11 @@ fn main() {
 				100 if charging => ("", "#50FF00"),
 				_ => ("", "#50FF00"),
 			};
-			bfmt![
+			Ok(bfmt![
 				fg[color]
 				fmt["{}{} {:.0}%", if charging { "" } else { "" }, icon, capacity]
-			]
-		}))
+			])
+		})))
 		// Brightness
 		.register_fn("bright_up", || Backlight::adjust(0.05).unwrap_or(()))
 		.register_fn("bright_down", || Backlight::adjust(-0.05).unwrap_or(()))
@@ -353,4 +254,5 @@ fn main() {
 		.add(Text::new(bfmt![text["(◕ᴗ◕✿)"]]))
 		.run();
 	//libnotify::uninit();
+Ok(())
 }
